+++ weblate/vcs/git.py
@@ -30,6 +30,7 @@ from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy
 from git.config import GitConfigParser
 
+from weblate.utils.render import render_template
 from weblate.utils.xml import parse_xml
 from weblate.vcs.base import Repository, RepositoryException
 from weblate.vcs.gpg import get_gpg_sign_key
@@ -576,6 +577,9 @@ class GitMergeRequestBase(GitRepository):
     def create_pull_request(self, origin_branch, fork_branch):
         raise NotImplementedError()
 
+    def get_merge_message(self):
+        return render_template(settings.DEFAULT_PULL_MESSAGE, component=self)
+
 
 class GithubRepository(GitMergeRequestBase):
 
@@ -610,7 +614,7 @@ class GithubRepository(GitMergeRequestBase):
                 '--base',
                 origin_branch,
                 '--message',
-                settings.DEFAULT_PULL_MESSAGE,
+                self.get_merge_message(),
             ]
         )
 
@@ -737,7 +741,7 @@ class GitLabRepository(GitMergeRequestBase):
                 'origin',
                 origin_branch,
                 '--message',
-                settings.DEFAULT_PULL_MESSAGE,
+                self.get_merge_message(),
             ]
         )
         # Return to the previous checked out branch.